// src/app/enums/code.ts
export enum Code {
  Black = '§0',
  DarkBlue = '§1',
  DarkGreen = '§2',
  DarkAqua = '§3',
  DarkRed = '§4',
  DarkPurple = '§5',
  Gold = '§6',
  Gray = '§7',
  DarkGray = '§8',
  Blue = '§9',
  Green = '§a',
  Aqua = '§b',
  Red = '§c',
  LightPurple = '§d',
  Yellow = '§e',
  White = '§f',
  MinecoinGold = '§g',
  Bold = '§l',
  Italic = '§m',
  Underline = '§n',
  Strike = '§o',
  Obfuscated = '§k',
  Reset = '§r',
}